Separate namespaces pubsub. eth, parity, net
