+++ js/src/api/pubsub/net/index.js
@@ -13,4 +13,4 @@
 
 // You should have received a copy of the GNU General Public License
 // along with Parity.  If not, see <http://www.gnu.org/licenses/>.
-export default from './pubsub';
+export default from './net';
